+++ examples/react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # example-react
 
+## 0.0.20
+
+### Patch Changes
+
+- Updated dependencies
+  - pinch-zoom-native@0.0.19
+
 ## 0.0.19
 
 ### Patch Changes