+++ packages/pinch-zoom-native/CHANGELOG.md
@@ -1,5 +1,11 @@
 # pinch-zoom-native
 
+## 0.0.19
+
+### Patch Changes
+
+- doubletap
+
 ## 0.0.18
 
 ### Patch Changes